Update Remove-PFPermissionsFromSSV.ps1
I ran the script on a 2013 to Exchange Online migration.  I had to make these changes on Exchange 2013 to get the script to work with username as the .DisplayName property was needed and with Public Folder names that had a space in them as the original only returned part of the name.